TestSuite: fix format warnings, comparison warnings, parameter warnings

+++ TestSuite/avltree.cpp
@@ -43,7 +43,7 @@ int TestAVLTree_CString()
 	TEST_ASSERT(avltree->size() == 0);
 
 	/* Simplest sanity checks done, now create some random data */
-	for (size_t i = 0; i < TREE_ITEMS; i++)
+	for (unsigned int i = 0; i < TREE_ITEMS; i++)
 	{
 		TEST_ASSERT(strings[i] == NULL);
 		strings[i] = new char[20];
@@ -118,7 +118,7 @@ int TestAVLTree_String()
 	TEST_ASSERT(avltree->size() == 0);
 
 	/* Simplest sanity checks done, now create some random data */
-	for (size_t i = 0; i < TREE_ITEMS; i++)
+	for (unsigned int i = 0; i < TREE_ITEMS; i++)
 	{
 		char buf[20];
 		buf[0] = (char)0;
